Spare work - complete change of system
After working with this blog system for a while, I have discovered that it is not up to standards and has many design flaws. So I am going to rebuild it partially, with a new system, this is for the books in case I want to use part of this again.